using the HAMT for now as my terrible MMPT from WNFS v1 is buggy.  I'm using a fork b/c the HAMT API should really have a write method, and would _love_ it if `cid.CID` implemented the CBOR marshaling interface natively.

This is also showing performance improvements over even the buggy MMPT, I suspect b/c this one only writes when the root is updated. Benchmarks are comparing to master @ e087580fb7716a88a9a36a3d41e6a711a801ec17

```
benchmark                                                old ns/op     new ns/op     delta
BenchmarkIPFSCat10MbFile-4                               24099         21202         -12.02%
BenchmarkIPFSWrite10MbFile-4                             41901         38335         -8.51%
BenchmarkIPFSCat10MbFileSubdir-4                         41545         29462         -29.08%
BenchmarkIPFSWrite10MbFileSubdir-4                       83816         77827         -7.15%
BenchmarkIPFSCp10DirectoriesWithOne10MbFileEach-4        2032281       2069801       +1.85%
BenchmarkPublicCat10MbFile-4                             59938         28857         -51.86%
BenchmarkPublicWrite10MbFile-4                           334974        257696        -23.07%
BenchmarkPublicCat10MbFileSubdir-4                       98454         49662         -49.56%
BenchmarkPublicWrite10MbFileSubdir-4                     568033        419323        -26.18%
BenchmarkPublicCp10DirectoriesWithOne10MbFileEach-4      4707          4134          -12.17%
BenchmarkPrivateCat10MbFile-4                            46211         38446         -16.80%
BenchmarkPrivateWrite10MbFile-4                          3433741       3340183       -2.72%
BenchmarkPrivateCat10MbFileSubdir-4                      54347         53768         -1.07%
BenchmarkPrivateWrite10MbFileSubdir-4                    4140070       3866860       -6.60%
BenchmarkPrivateCp10DirectoriesWithOne10MbFileEach-4     2788          2689          -3.55%

benchmark                                                old allocs     new allocs     delta
BenchmarkIPFSCat10MbFile-4                               42             42             +0.00%
BenchmarkIPFSWrite10MbFile-4                             204            204            +0.00%
BenchmarkIPFSCat10MbFileSubdir-4                         123            123            +0.00%
BenchmarkIPFSWrite10MbFileSubdir-4                       524            524            +0.00%
BenchmarkIPFSCp10DirectoriesWithOne10MbFileEach-4        2473           2472           -0.04%
BenchmarkPublicCat10MbFile-4                             108            108            +0.00%
BenchmarkPublicWrite10MbFile-4                           1168           1186           +1.54%
refactor(HAMT): replace MMPT with go-hampt-ipld
BenchmarkPublicCat10MbFileSubdir-4                       243            243            +0.00%
BenchmarkPublicWrite10MbFileSubdir-4                     2029           2047           +0.89%
BenchmarkPublicCp10DirectoriesWithOne10MbFileEach-4      20             18             -10.00%
BenchmarkPrivateCat10MbFile-4                            93             93             +0.00%
BenchmarkPrivateWrite10MbFile-4                          1756           1276           -27.33%
BenchmarkPrivateCat10MbFileSubdir-4                      154            154            +0.00%
BenchmarkPrivateWrite10MbFileSubdir-4                    2555           1653           -35.30%
BenchmarkPrivateCp10DirectoriesWithOne10MbFileEach-4     10             9              -10.00%

benchmark                                                old bytes     new bytes     delta
BenchmarkIPFSCat10MbFile-4                               49793         49834         +0.08%
BenchmarkIPFSWrite10MbFile-4                             11445         11382         -0.55%
BenchmarkIPFSCat10MbFileSubdir-4                         53420         53419         -0.00%
BenchmarkIPFSWrite10MbFileSubdir-4                       26986         27150         +0.61%
BenchmarkIPFSCp10DirectoriesWithOne10MbFileEach-4        253011        253030        +0.01%
BenchmarkPublicCat10MbFile-4                             53589         53595         +0.01%
BenchmarkPublicWrite10MbFile-4                           67518         68305         +1.17%
BenchmarkPublicCat10MbFileSubdir-4                       62011         61997         -0.02%
BenchmarkPublicWrite10MbFileSubdir-4                     117304        118048        +0.63%
BenchmarkPublicCp10DirectoriesWithOne10MbFileEach-4      1193          1095          -8.21%
BenchmarkPrivateCat10MbFile-4                            72976         73035         +0.08%
BenchmarkPrivateWrite10MbFile-4                          3053406       2870977       -5.97%
BenchmarkPrivateCat10MbFileSubdir-4                      80932         80902         -0.04%
BenchmarkPrivateWrite10MbFileSubdir-4                    4109331       3756259       -8.59%
BenchmarkPrivateCp10DirectoriesWithOne10MbFileEach-4     2471          2373          -3.97%
```